requirements.txt
beautifulsoup4==4.10.0
finnhub-python==2.4.6
matplotlib==3.5.0
nltk==3.6.5
numpy==1.21.4
pandas==1.3.4
python-dotenv==0.19.2
pytz==2021.3
regex==2021.11.10
requests==2.26.0
scikit-learn==1.0.1
seaborn==0.11.2
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@a6b6f3faaa26f541d9469651451340096b5abc92
tqdm==4.62.3
urllib3==1.26.7
yfinance==0.1.67
